pyproject.toml
[project]
name = "cv-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.115.0",
    "numpy>=2.4.4",
    "openai>=2.32.0",
    "pydantic-settings>=2.14.0",
    "sentence-transformers>=5.4.1",
    "uvicorn[standard]>=0.30.0",
]
